rename filenames to match component name
