+++ src/pages/landing/deriv-go/_lazy-load.js
@@ -1,5 +1,5 @@
 import Loadable from '@loadable/component'
 
-export const DerivGoContent = Loadable(() => import('./_deriv-go-profits'))
-export const StartDerivGo = Loadable(() => import('./_get-started'))
+export const DerivGoContent = Loadable(() => import('./_deriv-go-content'))
+export const StartDerivGo = Loadable(() => import('./_start-deriv-go'))
 export const FooterBanner = Loadable(() => import('./_footer-banner'))
