+++ src/pages/landing/deriv-go/components/_tabs.js
